Change to imported cases code
(A specified number par.nIndexCases)  of local cases at the start of the outbreak are now reclassified in main.m as imported cases to act as seed infections. Previously, the seed cases had to be saved as imported cases in the processed data files but now these are all saved as local for Ebola and the seeding assumption is handled in the main code. (Note situation is unchanged for Covid, where cases are already classified as local or imported in the data)